requirements.txt
astrapy==0.0.2
cassandra-driver==3.25.0
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
geomet==0.2.1.post1
idna==3.2
importlib-metadata==4.6.1
numpy==1.19.3
requests==2.26.0
simplejson==3.17.2
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.6
zipp==3.5.0
PyYAML~=5.4.1